fix: cross compile (#2)
