+++ .github/workflows/goreleaser.yml
@@ -1,50 +1,128 @@
-name: goreleaser
+name: Release
 
 on:
   push:
     tags:
       - 'v*'
+  workflow_dispatch:
 
 permissions:
   contents: write
   packages: write
 
 jobs:
-  build-linux:
-    runs-on: ubuntu-latest
+  build-python-backend:
+    name: Build Python Backend (${{ matrix.target_os }})
+    runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        goarch: [amd64]
+        include:
+          - os: ubuntu-latest
+            target_os: linux
+            binary_name: python-linux
+          - os: macos-latest
+            target_os: darwin
+            binary_name: python-darwin
+          - os: windows-latest
+            target_os: windows
+            binary_name: python-windows.exe
 
     steps:
-      - name: Checkout
-        uses: actions/checkout@v6
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.13'
+
+      - name: Set up uv
+        uses: astral-sh/setup-uv@v5
+
+      - name: Install dependencies
+        run: uv sync
+
+      - name: Generate Protobuf files (Python)
+        shell: bash
+        run: |
+          mkdir -p grpc_server/gen
+          touch grpc_server/gen/__init__.py
+          uv run python -m grpc_tools.protoc \
+            -Iproto \
+            --python_out=grpc_server/gen \
+            --grpc_python_out=grpc_server/gen \
+            --pyi_out=grpc_server/gen \
+            proto/music.proto
+          python -c "p='grpc_server/gen/music_pb2_grpc.py'; content=open(p).read(); open(p,'w').write(content.replace('import music_pb2 as', 'from . import music_pb2 as'))"
+
+      - name: Build PyInstaller executable
+        run: uv run pyinstaller main.spec
+
+      - name: Copy binary to backend directory
+        shell: bash
+        run: |
+          mkdir -p backend
+          if [ "${{ matrix.target_os }}" = "windows" ]; then
+            cp dist/main.exe backend/python-windows.exe
+          else
+            cp dist/main backend/${{ matrix.binary_name }}
+          fi
+
+      - name: Upload Python backend artifact
+        uses: actions/upload-artifact@v4
+        with:
+          name: python-backend-${{ matrix.target_os }}
+          path: backend/${{ matrix.binary_name }}
+          if-no-files-found: error
+
+  release:
+    name: Release Go Binaries via GoReleaser
+    needs: build-python-backend
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
         with:
           fetch-depth: 0
 
       - name: Set up Go
-        uses: actions/setup-go@v7
+        uses: actions/setup-go@v5
         with:
-          go-version: 1.26.x
+          go-version: '1.26.x'
+
+      - name: Download Linux Python backend artifact
+        uses: actions/download-artifact@v4
+        with:
+          name: python-backend-linux
+          path: backend/
+
+      - name: Download macOS Python backend artifact
+        uses: actions/download-artifact@v4
+        with:
+          name: python-backend-darwin
+          path: backend/
+
+      - name: Download Windows Python backend artifact
+        uses: actions/download-artifact@v4
+        with:
+          name: python-backend-windows
+          path: backend/
+
+      - name: Set permissions on Python backend executables
+        run: |
+          chmod +x backend/python-linux backend/python-darwin
+          ls -la backend/
 
       - name: Install ALSA & cross-compile dependencies
         run: |
           sudo apt-get update
           sudo apt-get install -y pkg-config libasound2-dev gcc-aarch64-linux-gnu libc6-dev-arm64-cross
 
-
       - name: Run GoReleaser
-        uses: goreleaser/goreleaser-action@v7
+        uses: goreleaser/goreleaser-action@v6
         with:
           version: '~> v2'
           args: release --clean
         env:
           CGO_ENABLED: 1
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: Upload artifacts
-        uses: actions/upload-artifact@v7
-        with:
-          name: linux-${{ matrix.goarch }}-artifacts
-          path: dist/
-
