+++ .gitignore
@@ -28,4 +28,7 @@ wheels/
 browser.json
 gen/
 backend/main
+backend/python-linux
+backend/python-darwin
+backend/python-windows.exe
 main.spec
